[run_container.h] Now you can define the name and the docker image with two arguments.

+++ run_container.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
-echo "Name:" $1
+echo "Container name:" $1
+echo "Image:" $2
 # Argument validation check
 if [ $# -eq 0 ]; then
     echo "Error: No arguments provided."
-    echo "Example: sh run_container.sh ubuntu_24_dqrobotics"
+    echo "Example: sh run_container.sh desired_container_name  docker_image"
     exit 1
 fi
+
+if [ $# != 2 ]; then
+    echo "Error: You need to provide two arguments."
+    echo "Example: sh run_container.sh desired_container_name  docker_image"
+    exit 1
+fi
+
 xhost +local:root
 docker run -it --name=$1 \
 --rm --privileged --network=host --env=DISPLAY \
-$1 /bin/bash
+$2 /bin/bash
